refactor(examples): move EHR example out of core, use registry executors
Addresses review on #2458: example code no longer ships in the oumi package.

- Move src/oumi/examples/ehr/ -> configs/examples/synthesis/ehr/, packaging
  executors + config in one directory next to each other.
- Collapse the per-tool executor files + _state helpers into a single
  self-contained executors.py (easier to reason about than scattered files).
- Executors register ehr.* names via @register_tool_executor; the config
  references those names, loaded through OUMI_EXTRA_DEPS_FILE (extra_deps.txt).
- Drop the executor unit tests.
